% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{function takes in a T/F statement
(stay with current selection or not), the door
that has already been opened, and the pick that
ends up being the final pick.}
}
\value{
The function returns a number between 1 and 3,
indicating the door that the contestant ended up
picking (could be the same as original or not).
}
\description{
This is the case where the constestant
is given the option to change their pick.
}
\details{
The contestant is given the option to stay with
their pick or change. The function runs through
two scenarios: the contestant stays with their
selection (stay) or not (!stay). If they stay,
then nothing changes and their original choice
remains (a.pick). If they do not stay, then their
new choice has to be the door that is not opened
(!= opened.door) and not chosen (!=a.pick).
}
\examples{
change_door( stay=F, opened.door=1, a.pick=3 )
}
