% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine If Contestant Has Won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{function takes in the final pick from
the previous step and the game.}
}
\value{
"WIN" or "LOSE"
}
\description{
This function will determine whether or not
the contestant wins (car selection)
}
\details{
In the game, if the contestant picks a
door with a goat behind it, they lose.
Conversely, if they pick a door with a car
behind it, they win. This function uses the
previous pick from the change door function.
So, this could either be the door the contestant
originally picked or the new one. This function
simply returns whether or not the contestant
won, based off of their door choice.
}
\examples{
determine_winner( final.pick=my.final.pick.stay, game=this.game )
}
