% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host Opens Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{function takes in a game and door pick.}
}
\value{
This function returns a number between 1 and 3
indicating a door with a goat behind it.
}
\description{
This function sets up the scenario where
the host will always open a door with a
goat behind it.
}
\details{
In this scenario, the host will always open a door
with a goat behind it. This door can't be the one
the contestant already selected. That is, the door
needs to not have a car and can't be a current
selection.
The two if statements go through the two scenarios
where the contestant's pick has a car or goat behind it.
}
\examples{
open_goat_door( this.game, my.initial.pick )
}
