% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play the Game}
\usage{
play_game()
}
\arguments{
\item{No}{arguments are used in this function.}
}
\value{
The Win/Lose outcomes of each strategy.
}
\description{
play_game() simulates an actual run of
the game and compares the outcomes of
each strategy.
}
\details{
This function combines multiple elements
from the previous steps. A game, first pick,
and opened door are created. Two scenarios are
created: the contestant either stayed or switched
their original door pick. The outcomes of each
scenario are created. The strategies are named,
outcomes are defined, and the function returns
which strategy yielded which result (WIN/LOSE).
}
\examples{
play_game()
}
