% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play 100 Games (with Outcomes)}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{This}{function takes in "n": number of games}
}
\value{
A percentage of wins/losses within the
100 plays.
}
\description{
play_n_games is a simulation that plays
n games (in this case, n = 100), and reports
the proportion of wins/losses.
}
\details{
Here, we get to see the game in action, multiple
times. The loop goes through the set number of games,
and then results.df returns the proportion of wins and
losses within those 100 plays.
}
\examples{
play_n_games(n=100)
results.df
}
