% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant Selects a Door}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used in this function.}
}
\value{
A number between 1 and 3, indicating which door was picked.
}
\description{
'select_door()' makes a random door selection between
1 and 3.
}
\details{
In the specified game, there are three doors with three
options: goat, goat, or car. This function gives the
doors values of 1, 2, and 3 and randomizes one pick
of the three doors.
}
\examples{
select_door()
}
