% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_goat_door.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Opens one of the doors with a goat behind it.}
\usage{
open_goat_door(a.game, a.pick)
}
\arguments{
\item{'a.game'}{and 'a.pick' are needed for the function to process. 'a.game' is a length 3 charactor vector
indicating the positions of the goats and cars. 'a.pick' is a numberic value between 1 and 3,
representing the contestants selected door.}
}
\value{
The function returns a numeric value between 1 and 3, representing the avalible door with a goat behind it, 'opened.door'.
}
\description{
'open_goat_door()' will open one of the two goat doors.
}
\details{
During this stage of the game show, the host will open one of the doors with a goat behind it.
This function will randomly open one of the two goat doors if the 'select_door()' function
retuned the door with the car. If the 'select_door()' function returned a door with a
goat, 'open_goat_door()' will return the other avalible goat door.
}
\examples{
open_goat_door( a.game, a.pick )

}
