% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_game.R
\name{play_game}
\alias{play_game}
\title{Create and execute a new Monty Hall Problem game.}
\usage{
play_game()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
The function will return a table based on the strategy outcome of stay and switch.
Each strategy outcome will return a length 1 character vector indicating either WIN' or 'LOSE'.
}
\description{
'play_game()' generates and executes a new game that consists of two doors with goats behind them, and one with a car.
}
\details{
This function will generate and execute a new game that tests
both strategies - staying with the originally selected door and switching
doors after a goat door has been revealed. The results of both strategies will be returned at the conclusion of the game.
}
\examples{
play_game()

}
