% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_n_games.R
\name{play_n_games}
\alias{play_n_games}
\title{Create and execute a new Monty Hall Problem game n times.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{'n'}{is needed for this function to process. 'n' is a positive, whole integer, 1 or greater.}
}
\value{
The function will return a table based on the outcome for the strategy stay and the
strategy switch with the proportions under 'LOSE' and 'WIN'. Additionally, the function
will return a table based on the strategy outcome of stay and switch. Each strategy
outcome will return a length 1 character vector indicating either WIN' or 'LOSE'.
The strategy outcome will be listed for each game played as defined by 'n'.
}
\description{
'play_n_games()' generates and executes a new game that consists of two doors with goats behind them, and one with a car n times.
}
\details{
This function will generate and execute a new game n time that tests
both strategies - staying with the originally selected door and switching
doors after a goat door has been revealed. The results of both strategies
will be returned at the conclusion of the game for n times.
}
\examples{
play_n_games( n=100 )

}
