% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_door.R
\name{select_door}
\alias{select_door}
\title{Randomly select one of three doors.}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
This function returns a numeric value between 1 and 3, representative of the contestants door choice, 'a.pick'.
}
\description{
'select_door()' will randomly select one of three doors and return a number between one and three.
}
\details{
This setup replicates what a contestant would do during this stage of the game show - select door one, two, or three.
}
\examples{
select_door()

}
